requirements.txt
pocket==0.3.5
nose==1.3.0
PyYAML==3.11